+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 1.6.3 Release (19.09.2020)
+* Updated locales to latest version
+* Added the possibility to change messages of locales on language.yml 
+
 ### 1.6.2 Release (15.08.2020 - 13.09.2020)
 * (Developer Alert - Changed class names)
 * Added support for 1.16 hex colors